refactor(role_service): move DTOs to web/dto
