+++ src/main/java/com/renewsim/backend/role_service/application/port/in/GetRolesUseCase.java
@@ -1,6 +1,6 @@
 package com.renewsim.backend.role_service.application.port.in;
 
-import com.renewsim.backend.role_service.dto.RoleDTO;
+import com.renewsim.backend.role_service.web.dto.RoleDTO;
 import com.renewsim.backend.shared.common.application.port.in.GetAllUseCase;
 
 public interface GetRolesUseCase extends GetAllUseCase<RoleDTO> {
