+++ src/main/java/com/renewsim/backend/user_service/infrastructure/client/RoleServiceClient.java
@@ -4,7 +4,7 @@ import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 
-import com.renewsim.backend.role_service.dto.RoleDTO;
+import com.renewsim.backend.role_service.web.dto.RoleDTO;
 import com.renewsim.backend.shared.dto.OperationResponse;
 
 @FeignClient(name = "role-service", contextId = "userRoleClient", url = "${role-service.url}", configuration = FeignRoleConfig.class)
