+++ src/main/java/com/renewsim/backend/user_service/infrastructure/persistence/adapter/RoleCatalogFeignAdapter.java
@@ -5,7 +5,7 @@ import java.util.Optional;
 import org.springframework.stereotype.Component;
 
 import com.renewsim.backend.shared.domain.vo.RoleName;
-import com.renewsim.backend.role_service.dto.RoleDTO;
+import com.renewsim.backend.role_service.web.dto.RoleDTO;
 import com.renewsim.backend.shared.dto.OperationResponse;
 import com.renewsim.backend.user_service.application.port.out.RoleCatalogPort;
 import com.renewsim.backend.user_service.web.dto.RoleSnapshot;
