+++ src/test/java/com/renewsim/backend/role_service/application/service/RoleServiceImplTest.java
@@ -6,7 +6,7 @@ import com.renewsim.backend.role_service.application.result.RoleCreationResultDT
 import com.renewsim.backend.role_service.domain.model.Role;
 import com.renewsim.backend.shared.domain.vo.RoleName;
 import com.renewsim.backend.role_service.domain.service.RoleDomainService;
-import com.renewsim.backend.role_service.dto.RoleDTO;
+import com.renewsim.backend.role_service.web.dto.RoleDTO;
 import com.renewsim.backend.role_service.application.mapper.RoleDtoMapper;
 import com.renewsim.backend.shared.exception.RoleAlreadyExistsException;
 import com.renewsim.backend.shared.exception.RoleNotFoundException;
